refactor: 잘못된 곳에 위치했던 클래스들 수정 (service -> domain)
