+++ src/test/java/roomescape/unit/reserveticket/ReserveTicketFinder.java
@@ -1,0 +1,10 @@
+package roomescape.unit.reserveticket;
+
+import roomescape.domain.reserveticket.ReserveTicket;
+
+class ReserveTicketFinder {
+
+    void reservation_id_로_member_를_찾을_수_있다() {
+        ReserveTicket reserveTicket
+    }
+}
