+++ src/test/java/roomescape/unit/service/reservationmember/ReserveTicketServiceTest.java
@@ -13,6 +13,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import roomescape.domain.reservation.ReservationStatus;
+import roomescape.domain.reserveticket.ReserveTicketWaiting;
 import roomescape.dto.member.SignupRequestDto;
 import roomescape.dto.reservation.AddReservationDto;
 import roomescape.dto.reservationtime.AddReservationTimeDto;
@@ -24,7 +25,6 @@ import roomescape.service.reservation.ReservationService;
 import roomescape.service.reservation.strategy.ReservationDuplicateCheckStrategy;
 import roomescape.service.reservationtime.ReservationTimeService;
 import roomescape.service.reserveticket.ReserveTicketService;
-import roomescape.service.reserveticket.ReserveTicketWaiting;
 import roomescape.service.theme.ThemeService;
 import roomescape.unit.config.ServiceFixture;
 import roomescape.unit.repository.member.FakeMemberRepository;
